+++ packages/widgetbook/test/helper/widget_test_helper.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+extension WidgetTesterExtension on WidgetTester {
+  Future<void> pumpWidgetWithMaterialApp(
+    Widget widget,
+  ) async {
+    return pumpWidget(
+      MaterialApp(
+        home: widget,
+      ),
+    );
+  }
+}
